Backport DataQuality and IncidentManager lazy-loading updates to 1.13 (#30126)
* feat(ui): split CustomizePageUtils.ts into 3 focused modules (#28937)

* feat(ui): split CustomizePageUtils.ts into 3 focused modules

Closes open-metadata/openmetadata-collate#4230

Splits 830-line CustomizePageUtils.ts (with 25 eager ClassBase imports)
into three focused files to eliminate unnecessary bundle weight on every
entity detail page:

- CustomizePageEntityTabUtils.ts  — tab display helpers (no ClassBase deps)
  sortTabs, getTabDisplayName, getTabLabelMapFromTabs,
  checkIfExpandViewSupported, getDetailsTabWithNewLabel

- CustomizePageDispatchUtils.ts   — dispatch/admin flow helpers (keeps all
  25 ClassBase singletons; only imported by admin/customize pages)
  getDefaultTabs, getDefaultWidgetForTab, getCustomizableWidgetByPage,
  getDummyDataByPage, getWidgetsFromKey, getWidgetHeight,
  getGlossaryTermDefaultTabs, getGlossaryDefaultTabs

- CustomizePageWidgetUtils.ts     — widget/layout helpers (no ClassBase deps)
  getAddWidgetHandler, updateWidgetHeightRecursively,
  getLayoutFromCustomizedPage, asyncNoop

CustomizePageUtils.ts deleted; all ~50 consumer files updated to import
directly from the correct split module. No re-export facade.

Test mocks updated: removed stale mock blocks on wrong paths (previously
mocking CustomizePageUtils which had no effect once consumers moved to
CustomizePagePureUtils); added correct mocks on CustomizePagePureUtils
for FileClassBase and DirectoryClassBase tests.


* fix checkstyle

---------


* feat(ui): lazy-load tab-level and heavy components with withSuspenseFallback (#28830)

* feat(ui): lazy-load tab-level and heavy components with withSuspenseFallback

Convert 105 component/util files from eager imports to React.lazy() +
withSuspenseFallback() for heavy components: SchemaEditor, ActivityFeedTab,
ContractTab, CustomPropertyTable, TaskTabNew, EntitySummaryPanel,
ModalWithMarkdownEditor, BlockEditor, CodeEditor, and others.

Lazy constants are placed after all traditional imports. Removed eager
component imports are replaced with import type for props interfaces where
applicable. Zero TypeScript errors introduced.


* fix failing test

* fix failing test

* improve performance

* update

* address comments

* fix failing e2e tests

* fix failing tests

* Fix failing tests

* fix failing test

* address comments

* fix ui checkstyle

* fix query page import

* fix checkstyle

---------


* feat(ui): extract pure utility functions and add lazy-loading infrastructure (#29061)

* feat(ui): extract pure utility functions and add lazy-loading infrastructure

Extracts pure (React-free) utility logic into dedicated *ClassBase.ts files for better code splitting and testability. Adds new utility files: ExploreFilterUtils.ts, GlossaryTreeUtils.ts, TagTransformUtils.ts, LazyTagComponents.tsx. Updates AppRouter to use lazy-loaded components with withSuspenseFallback for improved initial load performance. Adds test coverage for extracted utility functions.

Ref: open-metadata/openmetadata-collate#4230


* address comments

* fix: apply ui-checkstyle fixes for feat/lazy-util-pr1-new-utils-infra


* fix failing tests

* fix checkstyle

* address comments

* address comments

---------


* feat(ui): update DataQuality & IncidentManager components for lazy-load utility pattern (#29064)

* feat(ui): update DataQuality & IncidentManager components for lazy-load utility pattern

Updates DataQuality and IncidentManager components to import from extracted pure utility modules. Updates DataQualityUtils, TestCaseUtils imports. Updates dataQualityDashboardAPI for consistent utility usage.

Ref: open-metadata/openmetadata-collate#4230


* address comments

* fix checkstyle, tests and build

---------


* test(ui): fix data quality backport unit tests

---------

(cherry picked from commit a0692c3cf062e3e6bbbceda2296ae0f8e963c2ba)
