+++ openmetadata-ui/src/main/resources/ui/src/components/DataQuality/AddDataQualityTest/EditTestCaseModal.tsx
@@ -40,11 +40,9 @@ import {
 } from '../../../rest/testAPI';
 import { getEntityName } from '../../../utils/EntityNameUtils';
 import { getColumnNameFromEntityLink } from '../../../utils/EntityPureUtils';
-import { getEntityFQN } from '../../../utils/FeedUtils';
-import {
-  generateFormFields,
-  getPopupContainer,
-} from '../../../utils/formUtils';
+import { getEntityFQN } from '../../../utils/FeedUtilsPure';
+import { getPopupContainer } from '../../../utils/formPureUtils';
+import { generateFormFields } from '../../../utils/formUtils';
 import { getNameFromFQN } from '../../../utils/FqnUtils';
 import { isValidJSONString } from '../../../utils/StringUtils';
 import { getTagsWithoutTier, getTierTags } from '../../../utils/TablePureUtils';
@@ -54,7 +52,6 @@ import { EntityAttachmentProvider } from '../../common/EntityDescription/EntityA
 import Loader from '../../common/Loader/Loader';
 import { EditTestCaseModalProps } from './AddDataQualityTest.interface';
 import ParameterForm from './components/ParameterForm';
-
 const EditTestCaseModal: React.FC<EditTestCaseModalProps> = ({
   visible,
   testCase,
