+++ openmetadata-ui/src/main/resources/ui/src/components/DataQuality/AddDataQualityTest/TestSuiteIngestion.tsx
@@ -40,11 +40,11 @@ import {
   deployIngestionPipelineById,
   updateIngestionPipeline,
 } from '../../../rest/ingestionPipelineAPI';
+import { getScheduleOptionsFromSchedules } from '../../../utils/CronExpressionUtils';
 import { getEntityName } from '../../../utils/EntityNameUtils';
 import { getNameFromFQN } from '../../../utils/FqnUtils';
 import { Transi18next } from '../../../utils/i18next/LocalUtil';
-import { getScheduleOptionsFromSchedules } from '../../../utils/SchedularUtils';
-import { getIngestionName } from '../../../utils/ServiceUtils';
+import { getIngestionName } from '../../../utils/ServicePureUtils';
 import {
   generateUUID,
   replaceAllSpacialCharWith_,
