+++ openmetadata-ui/src/main/resources/ui/src/components/DataQuality/ChartWidgets/TestCaseStatusPieChartWidget/TestCaseStatusPieChartWidget.test.tsx
@@ -37,6 +37,9 @@ jest.mock('../../../../utils/DataQuality/DataQualityUtils', () => ({
     total: 8,
   }),
   getPieChartLabel: jest.fn().mockReturnValue(<div>Test Label</div>),
+}));
+
+jest.mock('../../../../utils/DataQuality/DataQualityPureUtils', () => ({
   getTestCaseTabPath: jest.fn((status: TestCaseStatus) => ({
     pathname: '/data-quality/test-cases',
     search: `testCaseStatus=${status}`,
@@ -136,7 +139,7 @@ describe('TestCaseStatusPieChartWidget', () => {
 
   it('should pass onSegmentClick to CustomPieChart and navigate on segment click', async () => {
     const { getTestCaseTabPath } = jest.requireMock(
-      '../../../../utils/DataQuality/DataQualityUtils'
+      '../../../../utils/DataQuality/DataQualityPureUtils'
     ) as { getTestCaseTabPath: jest.Mock };
     const mockNavigate = (
       jest.requireMock('react-router-dom') as {
