+++ openmetadata-ui/src/main/resources/ui/src/components/DataQuality/IncidentManager/IncidentManagerPageHeader/IncidentManagerPageHeader.component.tsx
@@ -44,11 +44,11 @@ import {
   updateTestCaseIncidentById,
 } from '../../../../rest/incidentManagerAPI';
 import { updateTestCaseById } from '../../../../rest/testAPI';
-import { getNameFromFQN } from '../../../../utils/CommonUtils';
 import { getEntityName } from '../../../../utils/EntityNameUtils';
 import { getColumnNameFromEntityLink } from '../../../../utils/EntityPureUtils';
-import { getCommonExtraInfoForVersionDetails } from '../../../../utils/EntityVersionUtils';
-import { getEntityFQN } from '../../../../utils/FeedUtils';
+import { getCommonExtraInfoForVersionDetails } from '../../../../utils/EntityVersionUtilsPure';
+import { getEntityFQN } from '../../../../utils/FeedUtilsPure';
+import { getNameFromFQN } from '../../../../utils/FqnUtils';
 import { getPrioritizedEditPermission } from '../../../../utils/PermissionsUtils';
 import { getEntityDetailsPath } from '../../../../utils/RouterUtils';
 import { getTaskDetailPath } from '../../../../utils/TasksUtils';
@@ -62,7 +62,6 @@ import Severity from '../Severity/Severity.component';
 import TestCaseIncidentManagerStatus from '../TestCaseStatus/TestCaseIncidentManagerStatus.component';
 import './incident-manager.less';
 import { IncidentManagerPageHeaderProps } from './IncidentManagerPageHeader.interface';
-
 const IncidentManagerPageHeader = ({
   onOwnerUpdate,
   fetchTaskCount,
