+++ openmetadata-ui/src/main/resources/ui/src/components/DataQuality/IncidentManager/SqlQueryTab/SqlQueryTab.component.tsx
@@ -23,15 +23,14 @@ import { useTestCaseStore } from '../../../../pages/IncidentManager/IncidentMana
 import {
   getChangedEntityNewValue,
   getChangedEntityOldValue,
-  getChangedEntityStatus,
   getDiffByFieldName,
-  getDiffDisplayValue,
-} from '../../../../utils/EntityVersionUtils';
+} from '../../../../utils/EntityDiffPureUtils';
+import { getDiffDisplayValue } from '../../../../utils/EntityDiffUtils';
+import { getChangedEntityStatus } from '../../../../utils/EntityVersionUtilsPure';
 import Loader from '../../../common/Loader/Loader';
 import QueryViewer from '../../../common/QueryViewer/QueryViewer.component';
 import '../TestCaseResultTab/test-case-result-tab.style.less';
 import AddSqlQueryFormModal from './AddSqlQueryFormModal/AddSqlQueryFormModal.component';
-
 const SqlQueryTab = () => {
   const { testCase, isLoading } = useTestCaseStore();
   const { version } = useParams<{ version: string }>();
