+++ openmetadata-ui/src/main/resources/ui/src/utils/DataQuality/DataQualityUtils.tsx
@@ -20,11 +20,8 @@ import {
   omit,
   omitBy,
   parseInt,
-  startCase,
-  uniqBy,
 } from 'lodash';
 import QueryString from 'qs';
-import { Surface } from 'recharts';
 import { ReactComponent as AccuracyIcon } from '../../assets/svg/ic-accuracy.svg';
 import { ReactComponent as ColumnIcon } from '../../assets/svg/ic-column.svg';
 import { ReactComponent as CompletenessIcon } from '../../assets/svg/ic-completeness.svg';
@@ -55,15 +52,12 @@ import {
   TestDataType,
   type TestDefinition,
 } from '../../generated/tests/testDefinition';
-import type { DataInsightChartTooltipProps } from '../../interface/data-insight.interface';
 import type { TableSearchSource } from '../../interface/search.interface';
 import {
   DataQualityPageTabs,
   type DataQualityDashboardChartFilters,
 } from '../../pages/DataQuality/DataQualityPage.interface';
 import type { ListTestCaseParamsBySearch } from '../../rest/testAPI';
-import { getEntryFormattedValue } from '../DataInsightUtils';
-import { formatDate } from '../date-time/DateTimeUtils';
 import EntityLink from '../EntityLink';
 import { getColumnNameFromEntityLink } from '../EntityPureUtils';
 import { getEntityFQN } from '../FeedUtils';
@@ -442,70 +436,6 @@ export const TEST_LEVEL_OPTIONS: SelectionOption[] = [
     icon: <ColumnIcon />,
   },
 ];
-
-export const CustomDQTooltip = (props: DataInsightChartTooltipProps) => {
-  const {
-    active,
-    dateTimeFormatter = formatDate,
-    isPercentage,
-    payload = [],
-    timeStampKey = 'timestampValue',
-    transformLabel = true,
-    valueFormatter,
-    displayDateInHeader = true,
-  } = props;
-
-  if (active && payload?.length) {
-    // we need to check if the xAxis is a date or not.
-    const timestamp = displayDateInHeader
-      ? dateTimeFormatter(payload[0].payload[timeStampKey] || 0)
-      : payload[0].payload[timeStampKey];
-
-    const payloadValue = uniqBy(payload, 'dataKey');
-
-    return (
-      <div className="tw:bg-primary tw:rounded-xl tw:border tw:border-border-secondary tw:shadow-md tw:p-2.5">
-        <p className="tw:m-0 tw:text-primary tw:font-medium tw:text-xs">
-          {timestamp}
-        </p>
-        <hr className="tw:border-primary tw:my-2 tw:border-dashed" />
-        <div className="tw:flex tw:flex-col tw:gap-1">
-          {payloadValue.map((entry, index) => {
-            const value = entry.value;
-
-            return (
-              <div
-                className="tw:flex tw:items-center tw:justify-between tw:gap-6 tw:pb-1 tw:text-sm"
-                key={`item-${index}`}>
-                <span className="tw:flex tw:items-center">
-                  <Surface
-                    className="tw:mr-2"
-                    height={14}
-                    version="1.1"
-                    width={4}>
-                    <rect fill={entry.color} height="14" rx="2" width="4" />
-                  </Surface>
-                  <span className="tw:text-tertiary tw:text-[11px]">
-                    {transformLabel
-                      ? startCase(entry.name ?? (entry.dataKey as string))
-                      : entry.name ?? (entry.dataKey as string)}
-                  </span>
-                </span>
-                <span className="tw:font-medium tw:text-primary tw:text-[11px]">
-                  {valueFormatter
-                    ? valueFormatter(value, entry.name ?? entry.dataKey)
-                    : getEntryFormattedValue(value, isPercentage)}
-                </span>
-              </div>
-            );
-          })}
-        </div>
-      </div>
-    );
-  }
-
-  return null;
-};
 
 export type TestCaseCountByStatus = {
   success: number;
